[refactor](remove unused code) remove runtime tuple structure
