+++ be/src/exec/base_scanner.cpp
@@ -24,7 +24,6 @@
 #include "runtime/descriptors.h"
 #include "runtime/raw_value.h"
 #include "runtime/runtime_state.h"
-#include "runtime/tuple.h"
 #include "vec/data_types/data_type_factory.hpp"
 
 namespace doris {
