+++ be/src/exec/es/es_scroll_parser.h
@@ -21,7 +21,6 @@
 
 #include "rapidjson/document.h"
 #include "runtime/descriptors.h"
-#include "runtime/tuple.h"
 #include "vec/core/block.h"
 
 namespace doris {
