+++ be/src/exec/olap_common.h
@@ -27,7 +27,7 @@
 
 #include "exec/olap_utils.h"
 #include "olap/olap_common.h"
-#include "olap/tuple.h"
+#include "olap/olap_tuple.h"
 #include "runtime/primitive_type.h"
 #include "runtime/type_limit.h"
 #include "vec/io/io_helper.h"
