+++ be/src/exec/schema_scanner.h
@@ -24,7 +24,6 @@
 #include "gen_cpp/Descriptors_types.h"
 #include "gen_cpp/Types_types.h"
 #include "runtime/mem_pool.h"
-#include "runtime/tuple.h"
 #include "vec/core/block.h"
 
 namespace doris {
