+++ be/src/olap/memtable.cpp
@@ -23,7 +23,6 @@
 #include "olap/schema.h"
 #include "olap/schema_change.h"
 #include "runtime/load_channel_mgr.h"
-#include "runtime/tuple.h"
 #include "util/doris_metrics.h"
 #include "vec/aggregate_functions/aggregate_function_reader.h"
 #include "vec/aggregate_functions/aggregate_function_simple_factory.h"
