+++ be/test/olap/delta_writer_test.cpp
@@ -36,7 +36,6 @@
 #include "runtime/descriptor_helper.h"
 #include "runtime/exec_env.h"
 #include "runtime/mem_pool.h"
-#include "runtime/tuple.h"
 #include "util/file_utils.h"
 
 namespace doris {
