+++ be/test/olap/engine_storage_migration_task_test.cpp
@@ -37,7 +37,6 @@
 #include "runtime/descriptor_helper.h"
 #include "runtime/exec_env.h"
 #include "runtime/mem_pool.h"
-#include "runtime/tuple.h"
 #include "util/file_utils.h"
 
 namespace doris {
