+++ be/test/vec/exprs/vexpr_test.cpp
@@ -31,7 +31,6 @@
 #include "runtime/memory/chunk_allocator.h"
 #include "runtime/primitive_type.h"
 #include "runtime/runtime_state.h"
-#include "runtime/tuple.h"
 #include "testutil/desc_tbl_builder.h"
 #include "vec/exprs/vliteral.h"
 #include "vec/runtime/vdatetime_value.h"
